+++ src/main/java/ec/edu/espe/arquitectura/escolastico/organizacionFisica/model/Edificio.java
@@ -36,7 +36,7 @@ public class Edificio implements Serializable {
     @Column(name = "posee_aulas", nullable = false, length = 1)
     private String poseeAulas;
     @Column(name = "pisos", nullable = false)
-    private short pisos;
+    private Integer pisos;
     @Column(name = "aud_usuario", nullable = false, length = 30)
     private String audUsuario;
     @Column(name = "aud_fecha", nullable = false)
@@ -45,7 +45,7 @@ public class Edificio implements Serializable {
     @Column(name = "aud_ip", nullable = false, length = 30)
     private String audIp;
     @Column(name = "version", nullable = false)
-    private int version;
+    private Integer version;
     @JoinColumn(name = "cod_sede", referencedColumnName = "cod_sede")
     @ManyToOne
     private Sede sede;
@@ -123,11 +123,11 @@ public class Edificio implements Serializable {
         this.poseeAulas = poseeAulas;
     }
 
-    public short getPisos() {
+    public Integer getPisos() {
         return pisos;
     }
 
-    public void setPisos(short pisos) {
+    public void setPisos(Integer pisos) {
         this.pisos = pisos;
     }
 
@@ -155,11 +155,11 @@ public class Edificio implements Serializable {
         this.audIp = audIp;
     }
 
-    public int getVersion() {
+    public Integer getVersion() {
         return version;
     }
 
-    public void setVersion(int version) {
+    public void setVersion(Integer version) {
         this.version = version;
     }
