+++ src/main/java/ec/edu/espe/arquitectura/escolastico/persona/model/FamiliarPersona.java
@@ -39,7 +39,7 @@ public class FamiliarPersona implements Serializable {
     @Column(name = "aud_ip", nullable = false, length = 30)
     private String audIp;
     @Column(name = "version", nullable = false)
-    private int version;
+    private Integer version;
     @JoinColumn(name = "cod_persona", referencedColumnName = "cod_persona", nullable = false, insertable = false, updatable = false)
     @ManyToOne(optional = false)
     private Persona persona;
@@ -51,7 +51,7 @@ public class FamiliarPersona implements Serializable {
         this.pk = perFamiliarPersonaPK;
     }
 
-    public FamiliarPersona(int codPersona, short secFamiliarPersona) {
+    public FamiliarPersona(Integer codPersona, Integer secFamiliarPersona) {
         this.pk = new FamiliarPersonaPK(codPersona, secFamiliarPersona);
     }
 
@@ -135,11 +135,11 @@ public class FamiliarPersona implements Serializable {
         this.audIp = audIp;
     }
 
-    public int getVersion() {
+    public Integer getVersion() {
         return version;
     }
 
-    public void setVersion(int version) {
+    public void setVersion(Integer version) {
         this.version = version;
     }
