+++ src/main/java/ec/edu/espe/arquitectura/escolastico/seguridad/model/Usuario.java
@@ -41,7 +41,7 @@ public class Usuario implements Serializable {
     @Temporal(TemporalType.DATE)
     private Date fechaCambioClave;
     @Column(name = "nro_intentos_fallidos", nullable = false)
-    private short nroIntentosFallidos;
+    private Integer nroIntentosFallidos;
     @Column(name = "fecha_ultima_sesion")
     @Temporal(TemporalType.TIMESTAMP)
     private Date fechaUltimaSesion;
@@ -53,7 +53,7 @@ public class Usuario implements Serializable {
     @Column(name = "aud_ip", nullable = false, length = 30)
     private String audIp;
     @Column(name = "version", nullable = false)
-    private int version;
+    private Integer version;
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "usuario")
     private List<UsuarioPerfil> perfiles;
 
@@ -136,11 +136,11 @@ public class Usuario implements Serializable {
         this.fechaCambioClave = fechaCambioClave;
     }
 
-    public short getNroIntentosFallidos() {
+    public Integer getNroIntentosFallidos() {
         return nroIntentosFallidos;
     }
 
-    public void setNroIntentosFallidos(short nroIntentosFallidos) {
+    public void setNroIntentosFallidos(Integer nroIntentosFallidos) {
         this.nroIntentosFallidos = nroIntentosFallidos;
     }
 
@@ -176,11 +176,11 @@ public class Usuario implements Serializable {
         this.audIp = audIp;
     }
 
-    public int getVersion() {
+    public Integer getVersion() {
         return version;
     }
 
-    public void setVersion(int version) {
+    public void setVersion(Integer version) {
         this.version = version;
     }
